JoinControl moves to components/ and owns its own prop contract (#2656)
A shared control cannot live under pages/factionDetail/ once a second,
non-page host mounts it, and moving the file while it still imported
Membership from that page's hook would have moved nothing. So the four
fields the trio actually reads become JoinTarget, exported beside the
control; Membership extends it and keeps the fifth (state), which the trio
never reads.

Also adds autoFocus, off by default: a modal host needs the freshly-mounted
affirmative focused so a short viewport scrolls it into view (#2130), and
an inline faction body must not grab focus.
