+++ frontend/src/pages/factionDetail/archetypes/CovenFactionBody.tsx
@@ -30,7 +30,7 @@ import {
 import { computeFactionMultiplier } from "../../../utils/points";
 import { factionName, factionDescription } from "../../../utils/factions";
 import type { CharacterOut } from "../../../api/auth";
-import { JoinControl, type JoinControlSkin } from "../JoinControl";
+import { JoinControl, type JoinControlSkin } from "../../../components/JoinControl";
 import { SectionPanel, SectionToggle, useFactionSections } from "../sectionDisclosure";
 import type { FactionDetailState } from "../useFactionDetail";
